split RBRemovableMediaSource up into device source and transfer target
RBRemovableMediaSource tied together too many unrelated aspects -
the source UI structure, the ability to eject the device, and the
ability to transfer things to the device.  In some cases (particularly
audio CDs) we only want some of those things.

RBDeviceSource is an interface for things that can be ejected.  The
default implementation operates on a GMount or a GVolume, but the
MTP plugin does its own thing.

RBTransferTarget is an interface for things that can accept transfers.
Currently only implemented by the generic player, iPod and MTP plugins,
but the library might also use it.  The implementation needs to provide
an encoding target and a method for building destination URIs.

The remaining parts of RBRemovableMediaSource now live in
RBMediaPlayerSource.