rework search UI and add source toolbar
This replaces the search toolbar with a menu accessible through the
primary icon in the search entry, and makes the search entry look
more like search entries elsewhere in GNOME, particularly evolution,
empathy and the control center.

In the same row as the search entry, we now have a toolbar containing
source-specific actions, such as adding new radio streams or podcasts,
and toggling the visibility of the browser.  Sources and plugins no
longer have the ability to add items to the main toolbar.  Since the
main toolbar layout is now static, we can look at changing to a more
compact layout.

There is now once toolbar instance per source, rather than one shared
between them.  The toolbar takes care of remembering search state
for the source it's attached to, rather than having it stored externally.