add extdb, a database for storing external metadata
extdb associates external metadata items, such as album art images,
with sets of properties and values, such as album and artist names.
It uses tdb to store the associations, and (for now) keeps the
actual data in separate files.

Lookups are based on keys that contain sets of required and optional
properties.  Required properties must be present in the item found,
and optional properties must match if present.  When making a request,
additional information can be provided, such as musicbrainz album IDs,
that will not be stored, but only used to facilitate searches.

When a request fails, extdb emits a signal to allow metadata providers
to perform a search and store the resulting data, which then answers
the request.

When new data is stored in an extdb instance, it emits a signal allowing
consumers to make use of it.

extdb will initially be used for album art, but will also store
lyrics and perhaps other data in the future.