xfade: fix race causing repeated attempts to link streams (bug #664831)
commit 38c658a introduced a window where we'd attempt to link a
stream to the adder twice, which would cause a "Failed to link new
stream into GStreamer pipeline" error on the second attempt.  Fix this
by adding a flag to the stream to indicate we're trying to link it,
protected by the stream list lock, so now we'll only try once.