magnatune: rework the loading screen (bug #667971)
Hello,

I worked with the owner of Magnatune, John Buckman, to create a new splash
screen for the Magnatune plugin. The current splash screen is very busy with
too much to read for the short time it is displayed.

The new screen pares down the bullet points to the ones that John thought were
the most important. I've tested the new one and it "works for me". I'm
attaching a diff as well as the entire replacement "ui" file.

  Mark