track-transfer-batch: avoid gtk_dialog_run for overwrite prompt
Replace the RBEncoder 'overwrite' signal with a parameter to
rb_encoder_encode, and just make it fail if overwrite is FALSE
and the destination file exists.

Change the transfer batch 'overwrite-prompt' signal to suspend
the batch and add a method to resume it, either overwriting the
existing destination file or skipping the track.

This means the overwrite-prompt signal handler doesn't have to
return the user's decision, so it doesn't have to call gtk_dialog_run.

Friends don't let friends gtk_dialog_run.