only load tracks from media player devices when first selected
This avoids lots of slow file access for generic players that
don't end up getting selected, and grabbing single-access MTP
devices when other things may be trying to talk to them.

We can now reliably select the device source via the
rhythmbox-device desktop file, so if the device mounter is
configured to invoke that, everything will work fine, whether
Rhythmbox is already running or not.