refstring: use g_strlcpy to avoid glibc braindamage
Certain versions of glibc have a broken SSE strcpy that writes
past the end of the destination.  This only seems to cause
problems on i386, where it results in random heap corruption
during startup, usually followed by crashes.  x86_64 seems to
be unaffected, probably due to a combination of malloc and
structure member alignment.

In comparison, (g_)strlcpy is a simple for loop, so it doesn't
crash.

I'm sure the optimised strcpy looks nice in benchmarks though.

https://bugzilla.gnome.org/show_bug.cgi?id=666066