new more compact header layout
Now we have the playback controls, the current track information
(now including album art), the time display (which can be toggled
between elapsed and remaining time), the position slider (much
shorter than before) and the volume control on the same line.

With this layout it makes no sense to hide the toolbar, so that
option is no longer available.  It also makes no sense to use
a toolbar style other than 'text below icons', so this option
has been removed too.

Small display mode is also gone.  It needs to be reimplemented
as a separate window.

This took me far too long land.