fading-image: try harder
Draw a border around the current image (or the next, if fading)
rather than around the widget's allocation.  This avoids ugly
gaps between the edge of the image and the border.

When already in the middle of a transition when starting another,
use the current transition state as the start image for the next,
rather than abruptly jumping to the next or reverting to the previous
image.

When starting a new transition, replace the current image with the
next, so current is updated even if the transition doesn't finish.