rhythmdb: limit the number of outstanding entry adds for an import job
Rather than feeding everything into the rhythmdb action queue
immediately, limit the number of URIs in flight,  This makes it
possible to cancel an import job and wait for any outstanding
operations to finish in a reasonable amount of time, rather than
minutes or more for large imports, and it doesn't appear to
impact overall time taken.  It may reduce peak memory usage too.