add a new dialog for importing music into the library
This replaces the previous menu items that just opened a file
chooser window allowing selection of a file or folder to import.
The new dialog scans all the files under a selected location and
presents them all in a track list, allowing individual tracks
to be added or copied into the library, or played without being
added at all.

In the future this will be extended to allow copying from media
players and network shares (daap or upnp; smb etc. are already
covered).

On initial startup (with no library locations set), the contents
of the default music directory are imported, and if that doesn't
add anything to the library, the import dialog is shown.