Restore Audioscrobbler radio source UI popup name
In commit 6fe21e1b the popup name for audioscrobbler radio sources was inexplicably changed in the UI XML but not in the code. I imagine it was changed by a find/replace gone wrong. That change broke the popup used to rename and delete radio stations. This commit reverts the XML update component of commit 6fe21e1b to re-enable the radio source context menu.

https://bugzilla.gnome.org/show_bug.cgi?id=679982