xfade: deal with tag messages posted by the sink
With GStreamer 1.0 tag messages are posted by the sink rather than
the element that found the tags, so we can't find the stream that
the tags apply to by checking the message source any more.  For now,
we associate tag messages with the first available stream that is
playing, but at some point we should be able to get a stream id
for the tags.