Port daap and magnatune plugins to libsecret
The Magnatune plugin currently falls back to not saving the password at all if
libsecret isn't found. This fallback code should be removed once libsecret
becomes as prevalent as gnome-keyring was.

Audioscrobbler used gnome-keyring for the old API, which no longer works.
Hence, that code was removed.

https://bugzilla.gnome.org/show_bug.cgi?id=694981

As a bonus, add cscope.out to .gitignore