Convert everything to GMenu and add an app menu
Most significantly, this removes the menu bar, scattering its contents
into an application menu, additional menus added to source toolbars,
and a toolbar in the source list.

The playlist related parts of the music menu are now in the source list
toolbar.  The rest formed the basis of the app menu.

Most sources now have an edit menu that corresponds roughly to the
edit menu from the menu bar.

The view and tools menus are now part of the app menu.

The control menu is gone, since it didn't do anything that wasn't already
represented in the controls in the main toolbar.

The help menu is now part of the app menu.