audioscrobbler: don't save submission queues without a username
Queues saved without a username are saved as
~/.local/share/rhythmbox/audioscrobbler/submission-queues/servicename
which prevents the creation of the directory used to store queues
saved with a username.  This data can't be submitted to the service,
so there's no point saving it.  Since it interferes with proper operation
later on, remove such a file if we find one on startup.