property-view: don't reset selection during property renames
If we've only seen the delete half of a property rename (such as
changing the genre of a radio stream where there are no other
streams with that genre), we can't reset the selection yet as
the property model is out of sync with its query model.  Use an
idle handler to defer this until the change has been fully
processed and everything is back in sync.

https://bugzilla.gnome.org/show_bug.cgi?id=692158