widget: reset elapsed time when song changes
This stops a brief flicker sometimes visible where
rb_header_playing_song_changed_cb is called before
rb_header_elapsed_changed_cb and so the duration of the song is
updated but the previous elapsed time still remains. It causes a
flicker of the incorrect time remaining and the progress slider to
briefly move to a new location.

https://bugzilla.gnome.org/show_bug.cgi?id=697959