add an interface for reporting task progress
RBTaskProgress consists of a few properties that can be used to
construct a list of tasks with progress bars for each.  Tasks
can also be cancellable.

RBTaskProgressSimple is a GObject that implements the interface.
It can be used to represent tasks that aren't bound to the lifecycle
of an existing object.