shell: fix window unmaximize behavior
When double clicking the title bar of the maximized main window, the
window is unmaximized and then instantaneously remaximized.

This is because at the end of rb_shell_window_state_cb() we call
rb_shell_sync_window_state() which read the window size saved values
from GSettings and reapplies them to the window. But because the
window was previously maximized, and the ConfigureNotify event hasn't
happen yet, the saved values are the ones from the maximized state.

https://bugzilla.gnome.org/show_bug.cgi?id=700424