button-bar: watch for model changes properly
The top level GMenu doesn't notify of changes to its nested children, so we
need to watch all of the GMenu instances that make up the button bar.  On
changes we disconnect and reconnect all the signal handlers.

The buttons and separators added to the grid need to be shown, otherwise items
added after the bar is shown initially don't appear.

Adding a separator at the start of the bar creates an ugly effect that sort of
looks like an extension of the GtkPaned handle, so we don't do that.