gst: Don't leak bus in RBEncoderGst
When starting the GstPipeline in RBEncoderGst, the associated GstBus
was not correctly unreffed. gst_pipeline_get_bus() returns a reference
to the bus object so it must be released after use. Most importantly,
as described in https://bugzilla.gnome.org/show_bug.cgi?id=683470 ,
we must make sure to remove the bus watch from the main loop
when we no longer need it, or this will cause the GstBus being watched
not to be properly disposed of.
In my testing, this was causing leaks of 2 fds per track transferred.
As the default max fd per process is 1024, this means that after about
500 tracks transferred, rhythmbox would get in big troubles as it
can no longer open new file descriptors.