import-dialog: redesign a bit
Now we have a single 'import' button that describes what will be
imported, with separate options (only one so far, whether to copy
files in to the library or not) determining what happens when it
gets clicked.  The remove button has been removed.

Closing the dialog now stops any scanning task that may have been
in progress.