Fix links to the website
The README file, as well as the « Rhythmbox Website » button inside the
application are still pointing to the old http://www.gnome.org/projects,
but everything was recently moved to the wiki.

I'm completely dropping the paragraph about development in the README,
because the information is already on the main wiki page.

https://bugzilla.gnome.org/show_bug.cgi?id=722069