grilo: add support for media in the root of grilo sources
When we finish browsing the root to find all the top level containers,
if we found some media in there, add an entry to the browser
representing the root, then select it in order to start browsing for media.

Since the root container is NULL, a few other changes were required to make
this work.  Marker rows are now represented as a different container type,
rather than a NULL container and we now pass the operation we want in
to start_media_browse rather than a NULL container means GRL_OP_SEARCH.

https://bugzilla.gnome.org/show_bug.cgi?id=727838