playbin: ensure tags don't persist into the next stream
During track changes, flush the bus to ensure that tag messages that
haven't made it to our bus handler yet are discarded.  This was most
noticeable when skipping tracks with embedded cover art, especially
when the next track didn't have it, as sometimes the cover art from
the previous track would be applied to the next.