rhythmdb: fix transfer annotation for rhythmdb_entry_new
rhythmdb_entry_new is a bit unusual in that the caller doesn't own a reference
to the newly created entry.  This isn't really a problem as the caller is
responsible for deleting its own entries (or at least any callers going via
introspection are), and absent any other refcounting bugs, the entry will
stay alive until it is deleted.

This fixes crashes when deactivating the magnatune plugin, or when quitting
with the magnatune catalog loaded.