Declared rb_metadata_iface_xml with extern
Because `rb-metadata-dbus.h` is included in multiple places
`rb_metadata_iface_xml` needs to be declared with `extern` in order for
Rhythmbox to build on Mac OS X with GCC 4.2.1.  (Apple's variant of GCC
may be more strict in this manner.)

It's probably best if it's declared with `extern` anyway since there
should only be one definition.