xfade: always attribute tags to the right stream
Tag messages are emitted by the sink, which makes it difficult to
associate them with the originating stream.  Previously we've just
assumed the stream that is currently in PLAYING state owns any tags
that come out, which isn't good enough.  Tag updates now include
the full tags from the stream, which means embedded cover art too.
Attributing this to the wrong stream means we get the wrong cover art
image.

To fix this, we inject a tag into each stream containing the stream uri,
which we use to locate the stream instance.  Injecting a tag into the
stream is a bit ugly but it effectively fixes the problem.

https://bugzilla.gnome.org/show_bug.cgi?id=743493