track-transfer-batch: use encoding settings
Encoding profile selection is now pretty complicated.  Now we use
the source format if possible, or the preferred format for the target,
or the first available lossy encoding, or any lossless encoding.

If the transcode-lossless setting is enabled, we'll never use a lossless
encoding, even if it's the source format.  The setting is not applied
if the preferred format is lossless.