Don't handle RTL icons manually for GTK+ >= 3.12
The filenames for icons in RTL languages has changed slightly after
GTK+ 3.12, including now the '-rtl' part after '-symbolic', and not
before as it used to be. Also, since that version on, GTK+ handles
this automatically, so there's no need to explicitly request the
RTL version of the icons files from the application.

This patch bumps the version GTK+ and forgets about manually requesting
the RTL versions of those icons since GTK+ will do that from now on.

https://bugzilla.gnome.org/show_bug.cgi?id=753359