rhythmdb: add metadata cache
This provides a mechanism for storing file metadata (not user metadata
such as ratings) outside of the normal rhythmdb file.  Initially it will
be used for files on media player devices, but it may find more uses.

This implementation stores metadata as GVariants in tdb files (much like
ext-db) but that also may change.  Cache entries also have a missing-since
field, which is used to purge entries that haven't been seen in some
period of time.