rhythmdb: add metadata cache helpers to RhythmDBEntryType
This lets an entry type specify a cache name (which maps to a tdb filename
in the current implementation) and methods for translating between entry
URIs and metadata cache keys.

Metadata cache keys are used because files on removable devices often don't
have a stable location, since the mount point for the device can change.
Metadata cache keys are constructed from some stable device identifier
(such as a device serial number) and the file path relative to the root of
the device.