media-player: add an entry type and implement metadata cache methods
Currently only the generic-player plugin can benefit from this, but
there willl be more in the future.

To make use of this, the media player class has to provide a URI prefix
(the mount point of the device) and a key prefix (the device identifier),
and call rb_media_player_source_purge_metadata_cache once all files from
the device have been processed.