rhythmdb: use metadata caches (via the entry type) during import
After reading metadata from a file, store it in the cache.  Ignore entries
may be cached too, since rereading them can take a significant amount of time.

After statting a file, attempt to load its metadata from the cache (if the
entry type has one); if metadata is found, compare the file size and
modification time with the stat results and discard on mismatches.

Ignore entries read from the cache are identified by their media type.
application/octet-stream will never be a valid audio media type, so it must
indicate an ignore entry, as errors are not cached.