file-helpers: improve msdos sanitization a bit
Everywhere rb_sanitize_path_for_msdos_filesystem is used, except for
rb_sanitize_uri_for_filesystem, it's used on individual components
of a path rather than the whole thing, so it makes sense for it to
also replace '/' with '-'.  rb_sanitize_uri_for_filesystem now calls
a different function with the old behaviour.