improve android device identification
Sadly it's not as easy as just looking for 'Android' in the device model string.
Anything with 'Android' or 'Nexus' in the model string, or 'motorola' or 'OnePlus'
in the vendor string is pretty safe to assume to be an Android device.
This covers most of the major vendors except Sony.  Most likely, more will be added
later.

Since the checks are more complicated now, they've been moved to the removable media
manager so both the android and mtp plugins can call them.