xfade: only try to set stream uri tags on decoders that handle it properly
The GstAudioDecoder base class doesn't send events downstream immediately,
which means we don't deadlock when we send events to it from a callback
that gets called because it sent an event downstream.  We can't be sure
that this is true for decoders that don't use this base class, so we play
it safe by now setting the stream uri tag on them.

https://bugzilla.gnome.org/show_bug.cgi?id=756992