android: tell user to unlock and enable MTP if no storage found
In Marshmallow, at least, you need to enable MTP each time the device is
connected.  Doing so makes the usb device disappear (which removes the existing
source) and reappear (creating a new one), so we don't need a button to trigger
a rescan here.  Perhaps this is different for older versions though.

https://bugzilla.gnome.org/show_bug.cgi?id=744775