xfade: don't crash when handling raw streams
The check to determine whether the audio decoder in use is derived
from GstAudioDecoder failed horribly on raw streams, where there's only
one level of ghost pad and then the audio source.  Checking to see if
the target of the first ghost pad is itself a ghost pad fixes the crash.

https://bugzilla.gnome.org/show_bug.cgi?id=762087