shell: Fix state of the "Party Mode" toggle item
The "Party Mode" menu item is a toggle menu item. If we only listen to
the "activate" signal, we'll listen for clicks, but we won't change the
state. So when the gear menu appears, the old state would be shown.

Listen to the state change instead, and update the action state.

http://bugzilla.gnome.org/show_bug.cgi?id=703910