build: Disable uninstalled build for srcdir != builddir
libpeas expects to find both the .plugin file for the plugin we're
trying to load, as well as the "binary" for the plugin in the
"module_dir" passed to peas_engine_add_search_path().

This isn't a problem for compiled plugins, as their binary will appear
in the hierarchy of the builddir, but Python plugins don't need to be
modified and the .plugin and the <modulename>.py files will end up in
different directories.

Disable uninstalled build support when srcdir != builddir to avoid
Python plugins not working in jhbuild. jhbuild defaults to srcdir !=
builddir, but installs the files anyway.