sources: Fix state of initial import not reflecting reality
When rhythmbox is started with an empty "locations" setting, it will
rewrite the setting with the user's music folder as the only location.
This will start an import job from the setting's "changed" signal. But
we already start another import job from the "load-complete" signal of
the database.

Make sure only the load-complete signal triggers an import when setting
the "locations" setting is empty on startup.