Replace Toolbar with Headerbar
Replace the toolbar with a headerbar, hereby unifying button sizes and modernizing some involved code. The album art no longer has a border in order to maximize its size in the headerbar. The tooltips for the play/pause/stop buttons are in the headerbar.ui file now, and therefore are marked translatable.

Fixes #1664.