removable-media-manager: monitor mount roots to work around gvfs bug
gvfs currently does not invalidate its mount info cache when a mount
is removed unless it has a connection open to the mount process, so
when another mount appears at the same path, it uses stale
information to talk to the mount process, so all operations fail.

We can work around this by monitoring the root of the mount, which
causes gvfs to keep a connection open.