artsearch: add support for coverartarchive.org
If we have a musicbrainz release ID, or we find one with the album/artist
search, try to fetch the 'front' image for the release from
coverartarchive.org.  If this fails, try the Amazon ASID too.

Since we don't know if coverartarchive.org has an image for a release without
making a http request, we now try to fetch the image and provide that to
ext-db, rather than giving it the url.