android: improve error handling in the scan process a bit
If the source is selected multiple times while still mounting the
volume, bail out rather than trying to scan for music files
repeatedly.

In callbacks for the operations used to look for music directories,
don't access the source until we've determined that the operation
wasn't cancelled - if it was, the source has been disposed, so
it can't be accessed.