podcast: rewrite downloader using libsoup directly rather than gvfs
This simplifies some parts of the implementation, since libsoup
lets us directly access the HTTP headers that we were trying to get
at through the gvfs layer.

It also makes it simple enough to retry the download (resuming from
where we got up to if possible) if we get an error that is likely to
be transient while reading from the input stream.