podcast: consider feed as parsed only on parse completion
In the old design, we had a separate channel for parsing the selected
feed in search results. This channel eventually replaced the search
model channel on parse completion. In the new design, parsing code and
feed model use a shared channel, which causes the feed model channel
to have partial 'posts', when parsing is in progress. So, rather than
using 'posts' field, we use the parse 'status' field to check whether
parsing has been completed.

Closes: #1849