podcast: use more appropriate error message and title for empty feeds
Currently we display "Unable to load the feed. Check your network
connection." in info bar for feeds with no items. Instead, we now
display a more appropriate "The feed does not contain any downloadable
items" ( translated ) error message in info bar.

Also, currently we display the feed URL as title in search results for
feeds with no items, when in fact a valid feed title might be
available. So, we use the title rather than the URL in that case.

Closes: #1690