encoder: fix missing unref on gtask object in some code paths
We are not unreffing gtask in some code paths. Here, we do it
immediately after g_task_run_in_thread (), as the function takes ref
to the task and unrefs it when the task is completed.

This was causing the system error "too many open files", as the
encoder was not getting unreffed, causing the pipeline not getting
unreffed, causing the gstreamer poll loop not getting unreffed,
causing the socketpair() used by the poll loop to not close. So, for
each track transfer, 2 fds were lost, causing the system error when
the open file count per process crossed the user ulimit of 1024 (
which normally happens when transferring more than 350 tracks ).