xfade: stop the sink when there's nothing playing (again)
This code has been here, #if 0'd out, for a long time, because it caused
problems with timestamps when the sink was started again.  Since we're
doing some stupid tricks with element states here, the base time of the
sink gets out of sync with the clock's time offset (they normally cancel
each other out), which caused audio to be dropped until the playback
time caught up with the time offset.

The solution then is to manually sync up the base time with the sink
after setting the pipeline back to PLAYING.  If the sink is a bin (it
always is), we recursively set the base time on the bin contents to be
sure we get it to the right place.