shell: Fix locale initialisation
Call setlocale() with "" as the locale to initialise locale support from
environment variables, rather than with NULL, which will return the
current locale setting in a retval we ignore.

This call has no effect on Rhythmbox when it works properly, as the
GApplication class code will call it, but as we need to call
setlocale() as early as possible, in case of initialisation problems.