xfade: try not to rely on position queries against stream bins
Position queries against stream bins are handled by decoder and
demuxer elements, so might be subject to bugs and other unfortunate
behaviour depending on the format and whether it's in push or pull
mode.  In particular, oggdemux in push mode will seek to the end of
the stream in order to work out the stream duration, and if queried
before it seeks back to the start, will return a position at the end
of thestream.  Depending when this happens, it can cause us to think
the stream is ending and skip to the next track, or use the wrong
time for the volume control when fading in the track.

To avoid having to do position queries, we have a couple of tricks.
When blocking a stream in order to unlink it, we record the timestamp
on the buffer that we get in the block callback, so we have a position
ready for when we want to relink the stream.  When we get a segment
event for a stream, we can use the start position in the segment.

If we're trying to fade a stream but we don't have a position yet,
defer the fade until we manage to adjust the stream base time.