xfade: handle linking and unblocking streams better
If the stream already has an adder pad, contine on with the rest
of the process anyway.  It doesn't really matter whether the
stream is blocked or not, we need to remove the block either way.
Sometimes a stream can finish buffering before it gets blocked,
which this function needs to be able to handle.