podcast: use guids to match up episodes with database entries
When updating a feed, we now query the list of existing entries from the
database to match up with episodes from the parsed feed, based on either
the source URL or GUID matching.  As before, database entries for
episodes no longer in the feed are deleted unless the episode has been
downloaded.