podcast: use track numbers to record the order of episodes in rss feeds
This improves use of feeds where multiple episodes are published with the
same publication date in a specific order.  Previously episodes with the
same date would be sorted by title, which would generally not maintain
the order from the feed.

Episodes that are no longer in the feed but still exist because they
have been downloaded have track numbers assigned in sequence starting
after the last episode in the feed, so ordering is preserved there too.