pythonconsole: add debugpy support
This is a different mechanism for debuggers to attach to the process,
so it gets a separate entry in the tools menu.

While here, make debugging a bit more discoverable.  The menu entries
are always added.  Selecting them starts the debug listener if possible,
and otherwise brings up an error box indicating which modules are
missing.