Don't explicitly specify icon sizes
Currently, icon sizes for the play controls are explicitly
specified in two places:

   (a) some are set to GTK_ICON_SIZE_DIALOG==6 (48px) in the UI
       definition files, and

   (b) some are overridden to GTK_ICON_SIZE_LARGE_TOOLBAR==4
       (24px) in the `construct_load_ui()` function.

However, all of the icons in question have the pixel-size
property explicitly specified as 24px in the UI definition file
as well. This property takes precedence over the icon_size
property when computing icon size for load and display, so omit
the icon-size property in both places.