Fix Last.fm top artists web request
While the Last.fm API does have a "library.getArtists" method, a
better one appears to be "user.getTopArtists". The artist list
returned by this call is windowed to the user-default period set
in the Last.fm account settings, and sorted descending.

Fix the JSON parser to parse the resulting response.