podcast: try to deal with episodes with the same filename
Before setting the download location for an episode, check if
another episode is already there.  If so, add a numeric suffix
(incrementing until an unused filename is found) to disambiguate.

It's technically possible that the numeric suffix could encounter
integer wraparound, but only if you download every episode of a
podcast that publishes one every second for 68 years, and they
all have the same filename.