build: Use get_variable function instead of running pkg-config
Running the command 'pkg-config' manually causes problems with distro
packaging or cross-compiling, since it might not run the same pkg-config
tool as the dependency was found with for building. Helpfully, meson
provides a builtin function to query variables of dependencies.

Note that I'm looking up the 'targets' variable on the gtk+-3.0
dependency instead of on gdk-3.0; this shouldn't cause any issue since
the 'targets' variable is available on both.