podcast: implement create_ext_db_key for podcast entry types
For feeds, add the feed URL as a field.  For the keys we create here
to match against entries, we have to use the entry tag name, subtitle.
For episodes, add the feed URL and also the episode guid as an optional
field, so we can match episode specific data if it exists, and fall back
to feed data otherwise.

Change the key used to store feed images to match the new lookup keys.