removable-media-manager: rework Android detection
Since there are very few new MTP devices being made that aren't
using the Android MTP implementation (but not none), it makes sense
to assume any MTP device we see is Android based unless we know
otherwise.  To do this, if built with libmtp, we only treat a device
as non-Android if it matches an entry in the libmtp device list
that doesn't have the set of flags applied to all Android devices.
This means we don't need to keep a manually updated list of Android
vendor and device names.