android: set the source's load status to 'loading' just before loading
While the source's load status is loading or waiting, the spinner on
the floating status bar is active, which takes up a noticeable amount
of cpu time even if it's not visible.

If we can't load anything because the android device isn't in file
transfer mode, still set load status to 'loaded' to deactivate the
spinner.

Fixes: #2112