multistream: add per-stream filter elements
This works the same way as for xfade - when constructing a new stream,
a signal is emitted that allows plugins to return GStreamer element
instances that get linked into the pipeline.  Filters are only added
when the pipeline is first constructed, and are never removed, so we
don't have to worry about modifying a running pipeline.